refactor: decompose __init__.py into cli_ui, delivery/, config/ submodules
Extract 1422 lines (53%) from the monolithic __init__.py (2692→1270 lines)
into focused submodules while maintaining full backward compatibility:

- cli_ui.py: StepTracker, interactive selectors, banner display
- delivery/agent_generator.py: 7 fallback agent content generators
- delivery/file_copier.py: file copy/install functions (agents, commands, hooks, etc.)
- config/settings.py: global/project permissions management
- config/mcp.py: MCP server configuration and .mcp.json handling

Quick wins included:
- Add BLUEPRINT_SCHEMA to schemas.py (blueprint.json was parsed without validation)
- Add validate_artifact() and load_and_validate() utilities to schemas.py
- Support jsonschema Draft7/4 fallback for older environments

All re-exports preserved in __init__.py — existing imports continue to work.
Tests updated: mock paths adjusted for new module locations.
24 new decomposition tests + 88 template sync tests pass.
